Using caching to speed up test suite on PRs. (#19)
We use ccache to cache the object files so that subsequent runs don't need to
recompile the whole shared library. Unfortunately, this doesn't work inside
tox, so we compromise by only running tox on master or tag pushes. For PRs, we
install and test manually, which allows us to use ccache; this speeds up the
acceptance of new features while retaining the clean check for master.

We also cache the pip-installed dependencies to save some init time.