Switched to using pybind11 (+ Cmake) with the latest tatami library. (#32)
This provides a more convenient interface between Python and C++, and simplifies
much of the build process (at the expense of requiring Cmake).

We remove the Mattress.h header to reduce downstream complexity. Now,
downstream code just needs to accept the shared pointer directly. We switch to
an unsigned int32 for the index to give us some more runway for large dims. 

Also updated the various github actions.