Do not accept array_t objects in the Python/C++ interface.
This avoids problems with inadvertent casting to get the expected type/layout,
which would often result in a dangling view in the initialized tatami::Matrix.
Rather, we accept array objects and manually check that they're of the expected
type/layout, just in case they weren't casted correctly on the Python side.