Revert to passing matrices to/from C++ via uintptr_t's. (#35)
This was initially motivated by the failure of pybind11's auto-conversion for
std::shared_ptr in downstream packages when using a prebuilt wheel of mattress.
More generally, this change avoids any reliance on pybind11 in downstream
packages; as long as they can take a uintptr_t, they can use mattress. 

We also introduce a BoundMatrix class inside a mattress.h header that can be
used by downstreams to get the casting correct. GC protection is now fully
handled in C++ via BoundMatrix::original, which reduces the risk of bugs when
memory management was previously split across C++ and Python.